+++ packages/vda-did/test/blockchain-api-mainnet-web3.test.ts
@@ -1,12 +1,13 @@
 const assert = require('assert')
 import { DID_LIST } from "@verida/vda-common-test";
 import BlockchainApi from "../src/blockchain/blockchainApi"
-import { VdaDidConfigurationOptions } from '@verida/types';
+import { BlockchainAnchor, VdaDidConfigurationOptions } from '@verida/types';
 import { Wallet } from "ethers";
 require('dotenv').config();
 
 // const did = Wallet.createRandom();
-const did = new Wallet(DID_LIST[0].privateKey);
+const didWallet = new Wallet(DID_LIST[0].privateKey);
+const did = `did:vda:${BlockchainAnchor.POLPOS}:${didWallet.address}`;
 
 const endPoints_A = ['https://A_1', 'https://A_2', 'https://A_3'];
 const endPoints_B = ['https://B_1', 'https://B_2'];
@@ -29,9 +30,9 @@ const configuration = {
 
 const createBlockchainAPI = (did: any) => {
     return new BlockchainApi(<VdaDidConfigurationOptions>{
-        identifier: did.address,
+        identifier: `did:vda:${BlockchainAnchor.POLPOS}:${did.address}`,
         signKey: did.privateKey,
-        chainNameOrId: "mainnet",
+        blockchain: BlockchainAnchor.POLPOS,
         ...configuration
     })
 }
@@ -39,7 +40,7 @@ const createBlockchainAPI = (did: any) => {
 describe('vda-did blockchain api', () => {
     let blockchainApi : BlockchainApi
     before(() => {
-        blockchainApi = createBlockchainAPI(did);
+        blockchainApi = createBlockchainAPI(didWallet);
     })
 
     describe('register', function() {
@@ -48,20 +49,20 @@ describe('vda-did blockchain api', () => {
         it('Register successfully', async () => {
             await blockchainApi.register(endPoints_A);
 
-            const lookupResult = await blockchainApi.lookup(did.address);
+            const lookupResult = await blockchainApi.lookup(did);
             assert.deepEqual(
                 lookupResult, 
-                {didController: did.address, endpoints: endPoints_A},
+                {didController: didWallet.address, endpoints: endPoints_A},
                 'Get same endpoints');
         })
 
         it('Should update for registered DID', async () => {
             await blockchainApi.register(endPoints_B);
 
-            const lookupResult = await blockchainApi.lookup(did.address);
+            const lookupResult = await blockchainApi.lookup(did);
             assert.deepEqual(
                 lookupResult, 
-                {didController: did.address, endpoints: endPoints_B}, 
+                {didController: didWallet.address, endpoints: endPoints_B}, 
                 'Get updated endpoints');
         })
 
@@ -74,7 +75,10 @@ describe('vda-did blockchain api', () => {
 
             await assert.rejects(
                 testAPI.register(endPoints_A),
-                {message: 'Failed to register endpoints'}
+                (err) => {
+                    assert.ok(err.message.startsWith('Failed to register endpoints'));
+                    return true;
+                }
             )
         })
     })