+++ packages/vda-did/test/blockchain-api-testnet-gasless.test.ts
@@ -1,10 +1,11 @@
 const assert = require('assert')
 import BlockchainApi from "../src/blockchain/blockchainApi"
 import { Wallet } from "ethers";
-import { VdaDidConfigurationOptions, VeridaWeb3ConfigurationOptions, Web3CallType } from '@verida/types';
+import { BlockchainAnchor, VdaDidConfigurationOptions, VeridaWeb3TransactionOptions, Web3CallType } from '@verida/types';
 require('dotenv').config();
 
-const did = Wallet.createRandom();
+const didWallet = Wallet.createRandom();
+const did = `did:vda:${BlockchainAnchor.POLAMOY}:${didWallet.address}`;
 
 const endPoints_A = ['https://A_1', 'https://A_2', 'https://A_3'];
 const endPoints_B = ['https://B_1', 'https://B_2'];
@@ -35,11 +36,11 @@ const configuration = {
     }
   }
 
-const createBlockchainAPI = (did: any) => {
+const createBlockchainAPI = (did: any, blockchain: BlockchainAnchor) => {
     return new BlockchainApi(<VdaDidConfigurationOptions>{
-        identifier: did.address,
+        identifier: `did:vda:${blockchain}:${did.address}`,
         signKey: did.privateKey,
-        chainNameOrId: "testnet",
+        blockchain,
         ...configuration
     })
 }
@@ -47,7 +48,7 @@ const createBlockchainAPI = (did: any) => {
 describe('vda-did blockchain api', () => {
     let blockchainApi : BlockchainApi
     before(() => {
-        blockchainApi = createBlockchainAPI(did);
+        blockchainApi = createBlockchainAPI(didWallet, BlockchainAnchor.POLAMOY);
     })
 
     describe('register', function() {
@@ -56,33 +57,36 @@ describe('vda-did blockchain api', () => {
         it.only('Register successfully', async () => {
             await blockchainApi.register(endPoints_A);
 
-            const lookupResult = await blockchainApi.lookup(did.address);
+            const lookupResult = await blockchainApi.lookup(did);
             assert.deepEqual(
                 lookupResult, 
-                {didController: did.address, endpoints: endPoints_A},
+                {didController: didWallet.address, endpoints: endPoints_A},
                 'Get same endpoints');
         })
 
         it('Should update for registered DID', async () => {
             await blockchainApi.register(endPoints_B);
 
-            const lookupResult = await blockchainApi.lookup(did.address);
+            const lookupResult = await blockchainApi.lookup(did);
             assert.deepEqual(
                 lookupResult, 
-                {didController: did.address, endpoints: endPoints_B}, 
+                {didController: didWallet.address, endpoints: endPoints_B}, 
                 'Get updated endpoints');
         })
 
         it('Should reject for revoked did', async () => {
             const tempDID = Wallet.createRandom();
-            const testAPI = createBlockchainAPI(tempDID)
+            const testAPI = createBlockchainAPI(tempDID, BlockchainAnchor.POLAMOY)
 
             await testAPI.register(endPoints_Empty);
             await testAPI.revoke();
 
             await assert.rejects(
                 testAPI.register(endPoints_A),
-                {message: 'Failed to register endpoints'}
+                (err) => {
+                    assert.ok(err.message.startsWith('Failed to register endpoints'));
+                    return true;
+                }
             )
         })
     })
@@ -90,32 +94,38 @@ describe('vda-did blockchain api', () => {
     describe('Lookup', function() {
         this.timeout(100 * 1000)
         it('Get endpoints successfully', async () => {
-            const lookupResult = await blockchainApi.lookup(did.address);
+            const lookupResult = await blockchainApi.lookup(did);
             assert.deepEqual(
                 lookupResult, 
-                {didController:did.address, endpoints:endPoints_B}, 
+                {didController:didWallet.address, endpoints:endPoints_B}, 
                 'Get updated endpoints');
         })
 
         it('Should reject for unregistered DID',async () => {
             const testDID = Wallet.createRandom();
-            const testAPI = createBlockchainAPI(testDID);
+            const testAPI = createBlockchainAPI(testDID, BlockchainAnchor.POLAMOY);
             await assert.rejects(
                 testAPI.lookup(testDID.address),
-                {message: 'DID not found'}
+                (err) => {
+                    assert.ok(err.message.startsWith('DID not found'));
+                    return true;
+                }
             )
         })
 
         it('Should reject for revoked DID', async () => {
             const testDID = Wallet.createRandom();
-            const testAPI = createBlockchainAPI(testDID);
+            const testAPI = createBlockchainAPI(testDID, BlockchainAnchor.POLAMOY);
 
             await testAPI.register(endPoints_A);
             await testAPI.revoke();
 
             await assert.rejects(
                 testAPI.lookup(testDID.address),
-                {message: 'DID not found'}
+                (err) => {
+                    assert.ok(err.message.startsWith('DID not found'));
+                    return true;
+                }
             )
         })
     })
@@ -124,16 +134,19 @@ describe('vda-did blockchain api', () => {
         this.timeout(100 * 1000)
         const controller = Wallet.createRandom();
         it('Should reject for unregistered DID', async () => {
-            const testAPI = createBlockchainAPI(Wallet.createRandom());
+            const testAPI = createBlockchainAPI(Wallet.createRandom(), BlockchainAnchor.POLAMOY);
             await assert.rejects(
                 testAPI.setController(controller.privateKey),
-                {message: 'Failed to set controller'}
+                (err) => {
+                    assert.ok(err.message.startsWith('Failed to set controller'));
+                    return true;
+                }
             )
         })
 
         it('Change controller successfully', async () => {
             const orgDID = Wallet.createRandom();
-            const testAPI = createBlockchainAPI(orgDID);
+            const testAPI = createBlockchainAPI(orgDID, BlockchainAnchor.POLAMOY);
 
             await testAPI.register(endPoints_Empty);
             const orgController = await testAPI.getController();
@@ -150,11 +163,14 @@ describe('vda-did blockchain api', () => {
 
     describe('Revoke', function() {
         this.timeout(100 * 1000)
-        const testAPI = createBlockchainAPI(Wallet.createRandom());
+        const testAPI = createBlockchainAPI(Wallet.createRandom(), BlockchainAnchor.POLAMOY);
         it('Should reject for unregistered DID', async () => {
             await assert.rejects(
                 testAPI.revoke(),
-                {message: 'Failed to revoke'}
+                (err) => {
+                    assert.ok(err.message.startsWith('Failed to revoke'));
+                    return true;
+                }
             );
         })
 
@@ -167,7 +183,10 @@ describe('vda-did blockchain api', () => {
         it('Should reject for revoked DID', async () => {
             await assert.rejects(
                 testAPI.revoke(),
-                {message: 'Failed to revoke'}
+                (err) => {
+                    assert.ok(err.message.startsWith('Failed to revoke'));
+                    return true;
+                }
             );
         })
     })