+++ packages/vda-reward-client/src/blockchain/userApi.ts
@@ -1,70 +1,20 @@
-import {
-    getVeridaContract,
-    VeridaContract
-} from "@verida/web3"
-import { Web3SelfTransactionConfig, VdaClientConfig, Network } from '@verida/types'
-import { ethers, Contract } from "ethers";
-import { getContractInfoForVeridaNetwork, RPC_URLS, getVeridaSign, getDefaultRpcUrl, DefaultNetworkBlockchainAnchors } from "@verida/vda-common";
-import { JsonRpcProvider } from '@ethersproject/providers';
-import { explodeDID } from '@verida/helpers'
+import { VdaClientConfig } from '@verida/types'
+import { ethers } from "ethers";
+import { getVeridaSign } from "@verida/vda-common";
+import { explodeDID } from '@verida/helpers';
+import { VeridaClientBase } from '@verida/vda-client-base';
 
 export interface ClaimType {
     reward: bigint,
     schema: string
 }
 
-export class VeridaRewardClient {
-
-    protected config: VdaClientConfig
-    protected network: Network
-    protected didAddress?: string
-
-    protected vdaWeb3Client? : VeridaContract
-
-    protected readOnly: boolean
-    protected contract?: ethers.Contract
+export class VeridaRewardClient extends VeridaClientBase {
 
     protected claimTypesCache: Record<string, ClaimType> = {}
 
     public constructor(config: VdaClientConfig) {
-        if (!config.callType) {
-            config.callType = 'web3'
-        }
-
-        this.config = config
-        this.readOnly = true
-        if (!config.web3Options) {
-            config.web3Options = {}
-        }
-
-        this.network = config.network
-
-        if (config.callType == 'web3' && !(<Web3SelfTransactionConfig>config.web3Options).rpcUrl) {
-            (<Web3SelfTransactionConfig> config.web3Options).rpcUrl = <string> RPC_URLS[this.network]
-        }
-
-        const contractInfo = getContractInfoForVeridaNetwork("VDARewardContract", this.network)
-        const blockchain = DefaultNetworkBlockchainAnchors[this.network]
-
-        if (config.did) {
-            this.readOnly = false
-            const { address } = explodeDID(config.did)
-            this.didAddress = address.toLowerCase()
-            
-            this.vdaWeb3Client = getVeridaContract(
-                config.callType, 
-                {...contractInfo,
-                ...config.web3Options})
-        } else {
-            let rpcUrl = (<Web3SelfTransactionConfig>config.web3Options).rpcUrl
-            if (!rpcUrl) {
-                rpcUrl = getDefaultRpcUrl(blockchain)!
-            }
-
-            const provider = new JsonRpcProvider(rpcUrl)
-
-            this.contract = new Contract(contractInfo.address, contractInfo.abi.abi, provider)
-        }
+        super(config, "reward");
     }
 
     /**
@@ -210,7 +160,8 @@ export class VeridaRewardClient {
             throw new Error(`Unable to submit to blockchain. No 'signKey' provided in config.`)
         }
 
-        let didAddress = did!.toLowerCase()
+        let didAddress = did!//.toLowerCase()
+        console.log("claim to storage : ",did, " : ", didAddress);
         if (didAddress.match('did')) {
             const { address } = explodeDID(did!)
             didAddress = address