+++ packages/vda-sbt-client/src/blockchain/blockchainApi.ts
@@ -1,65 +1,14 @@
-import {
-    getVeridaContract,
-    VeridaContract
-} from "@verida/web3"
-import { Web3SelfTransactionConfig, VdaClientConfig, Network } from '@verida/types'
-import { DefaultNetworkBlockchainAnchors, getContractInfoForVeridaNetwork, getDefaultRpcUrl } from "@verida/vda-common";
-import { JsonRpcProvider } from '@ethersproject/providers';
-import { explodeDID } from '@verida/helpers'
+import { VdaClientConfig } from '@verida/types'
 
-import { ethers, Contract, BigNumberish } from "ethers";
+import { ethers, BigNumberish } from "ethers";
 import EncryptionUtils from "@verida/encryption-utils";
+import { VeridaClientBase } from '@verida/vda-client-base';
 
 
-export class VeridaSBTClient {
-
-    private config: VdaClientConfig
-    private network: Network
-    private didAddress?: string
-
-    private readOnly: boolean
-    private vdaWeb3Client?: VeridaContract
-    private contract?: ethers.Contract
+export class VeridaSBTClient extends VeridaClientBase {
 
     public constructor(config: VdaClientConfig) {
-        if (!config.callType) {
-            config.callType = 'web3'
-        }
-
-        this.config = config
-        this.readOnly = true
-        if (!config.web3Options) {
-            config.web3Options = {}
-        }
-
-        this.network = config.network
-
-        const contractInfo = getContractInfoForVeridaNetwork("SoulboundNFT", this.network)
-        const blockchain = DefaultNetworkBlockchainAnchors[this.network]
-
-        if (config.callType == 'web3' && !(<Web3SelfTransactionConfig>config.web3Options).rpcUrl) {
-            (<Web3SelfTransactionConfig> config.web3Options).rpcUrl = getDefaultRpcUrl(blockchain)!
-        }
-
-        if (config.did) {
-            this.readOnly = false
-            const { address } = explodeDID(config.did)
-            this.didAddress = address.toLowerCase()
-            
-            this.vdaWeb3Client = getVeridaContract(
-                config.callType, 
-                {...contractInfo,
-                ...config.web3Options})
-        } else {
-            let rpcUrl = (<Web3SelfTransactionConfig>config.web3Options).rpcUrl
-            if (!rpcUrl) {
-                rpcUrl = getDefaultRpcUrl(blockchain)!
-            }
-
-            const provider = new JsonRpcProvider(rpcUrl)
-
-            this.contract = new Contract(contractInfo.address, contractInfo.abi.abi, provider)
-        }
+        super(config, "solboundNFT");
     }
 
     /**