+++ packages/vda-sbt-client/test/utils.ts
@@ -6,7 +6,7 @@ import {
 
 import { VeridaSBTClient } from "../src/index"
 import { Keyring } from "@verida/keyring";
-import { EnvironmentType } from "@verida/types";
+import { BlockchainAnchor } from "@verida/types";
 import { explodeDID } from "@verida/helpers"
 const assert = require('assert')
 
@@ -14,12 +14,13 @@ export const mintedTokenIds = [1, 3]
 export const burntTokenIds =  [2, 4]
 
 export const createBlockchainAPI = (did: DidInterface, isWallet = false) => {
+  const target_chain = BlockchainAnchor.DEVNET;
   const configuration = getBlockchainAPIConfiguration(did.privateKey);
   return new VeridaSBTClient({
-      did: isWallet ? 'did:vda:testnet:' + did.address : did.address,
-      signKey: did.privateKey,
-      network: EnvironmentType.TESTNET,
-      ...configuration
+    blockchainAnchor: target_chain,
+    did: isWallet ? `did:vda:${target_chain}:${did.address}` : did.address,
+    signKey: did.privateKey,
+    ...configuration
   })
 }