+++ packages/vda-reward-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+
+2024-06-28 (v)
+-------------------
+
+- Add a dependency to the `@verida/vda-client-base`
+- Updated test codes
+
 2024-02-09 (v3.0.0)
 -------------------